+++ Blackboard/Core/Formula/Actions/Define.cs
@@ -1,5 +1,4 @@
-﻿using Blackboard.Core.Extensions;
-using Blackboard.Core.Inspect;
+﻿using Blackboard.Core.Inspect;
 using Blackboard.Core.Nodes.Interfaces;
 using Blackboard.Core.Nodes.Outer;
 using Blackboard.Core.Types;
