+++ Blackboard/Core/Formula/Actions/TriggerGetter.cs
@@ -1,6 +1,7 @@
 ﻿using Blackboard.Core.Extensions;
 using Blackboard.Core.Inspect;
 using Blackboard.Core.Nodes.Interfaces;
+using Blackboard.Core.Record;
 using System.Collections.Generic;
 using System.Linq;
