+++ Blackboard/Core/Innate/Maker.cs
@@ -1,6 +1,5 @@
 ﻿using Blackboard.Core.Data.Caps;
 using Blackboard.Core.Data.Interfaces;
-using Blackboard.Core.Extensions;
 using Blackboard.Core.Formula.Actions;
 using Blackboard.Core.Inspect;
 using Blackboard.Core.Nodes.Inner;
