+++ Blackboard/Core/Nodes/Collections/NodeValueComparer.cs
@@ -1,5 +1,4 @@
-﻿using Blackboard.Core.Extensions;
-using Blackboard.Core.Nodes.Interfaces;
+﻿using Blackboard.Core.Nodes.Interfaces;
 using System.Collections.Generic;
 
 namespace Blackboard.Core.Nodes.Collections;
