+++ Blackboard/Core/Nodes/Inner/Format.cs
@@ -1,5 +1,4 @@
 ﻿using Blackboard.Core.Data.Caps;
-using Blackboard.Core.Extensions;
 using Blackboard.Core.Nodes.Bases;
 using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Functions;
