+++ Blackboard/Core/Nodes/Outer/ExternTrigger.cs
@@ -1,5 +1,4 @@
-﻿using Blackboard.Core.Extensions;
-using Blackboard.Core.Nodes.Bases;
+﻿using Blackboard.Core.Nodes.Bases;
 using Blackboard.Core.Nodes.Inner;
 using Blackboard.Core.Nodes.Interfaces;
