+++ Blackboard/Core/Types/Extensions.cs
@@ -1,13 +1,12 @@
 ﻿using Blackboard.Core.Nodes.Interfaces;
-using Blackboard.Core.Types;
 using System.Collections.Generic;
 using System.Linq;
 using S = System;
 
-namespace Blackboard.Core.Extensions;
+namespace Blackboard.Core.Types;
 
 /// <summary>The set of extensions for working with Blackboard Types.</summary>
-static public class TypeExt {
+static public class Extensions {
 
     /// <summary>The type nodes for all the given nodes.</summary>
     /// <param name="nodes">The nodes to get the types for.</param>
