+++ BlackboardTests/CoreTests/Benchmarks.cs
@@ -1,5 +1,4 @@
 ﻿using Blackboard.Core.Data.Caps;
-using Blackboard.Core.Extensions;
 using Blackboard.Core.Nodes.Functions;
 using Blackboard.Core.Nodes.Inner;
 using Blackboard.Core.Nodes.Outer;
