+++ BlackboardTests/ParserTests/Definitions.cs
@@ -1,6 +1,6 @@
 ﻿using Blackboard.Core;
-using Blackboard.Core.Extensions;
 using Blackboard.Core.Inspect;
+using Blackboard.Core.Record;
 using BlackboardTests.Tools;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
