+++ BlackboardTests/ParserTests/Literals.cs
@@ -1,5 +1,4 @@
 ﻿using Blackboard.Core;
-using Blackboard.Core.Extensions;
 using Blackboard.Core.Record;
 using BlackboardTests.Tools;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
