+++ BlackboardTests/ParserTests/Temps.cs
@@ -1,5 +1,5 @@
 ﻿using Blackboard.Core;
-using Blackboard.Core.Extensions;
+using Blackboard.Core.Record;
 using BlackboardTests.Tools;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
